backend-modified-segment-model

+++ backend/cadence/models/SegmentModel.py
@@ -3,9 +3,9 @@ from django.db import models;
 from django.core.validators import MinValueValidator;
 
 class Segment(models.Model):
-    distance = models.IntegerField(validators=[MinValueValidator(0)])
-    duration = models.IntegerField(validators=[MinValueValidator(0)])
-    pace = models.IntegerField(validators=[MinValueValidator(0)])
+    distance = models.FloatField(validators=[MinValueValidator(0)])
+    duration = models.FloatField(validators=[MinValueValidator(0)])
+    pace = models.FloatField(validators=[MinValueValidator(0)])
     run = models.ForeignKey(Run, on_delete=models.CASCADE)
     
     class Meta:
